checks classdefinitions on disk instead of in db
The InstallClassDefinition isInstalled method was checking the existance
of the objects in the database.

In some cases this caused Pimcore::initPlugins to take 90% of the total
request time.

We now check if the definition and the object files exist on disk and
could reduce the problematic requests to 50% of the original request
time.

Signed-off-by: Ike Devolder <ike.devolder@studioemma.eu>